+++ lib/src/models/models.dart
@@ -1,12 +1,14 @@
 // Base models
 export 'common.dart';
 export 'address.dart';
+export 'address_request.dart';
 export 'customer.dart';
 export 'order.dart';
 export 'cart.dart';
 export 'line_item.dart';
 export 'product.dart';
 export 'product_variant_option_value.dart';
+export 'product_variant_option_request.dart';
 export 'region.dart';
 export 'collection.dart';
 export 'category.dart';